feat(authentication)：Basic Information Management Delete Interface Add Administrator Authentication

+++ linkis-public-enhancements/linkis-basedata-manager/src/main/java/org/apache/linkis/basedatamanager/server/restful/GatewayAuthTokenRestfulApi.java
@@ -127,7 +127,12 @@ public class GatewayAuthTokenRestfulApi {
       httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(request, "Remove a Gateway Auth Token Record,id:" + id);
+    String username =
+        ModuleUserUtils.getOperationUser(
+            request, "Try to remove gateway auto token record with id:" + id);
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
     boolean result = gatewayAuthTokenService.removeById(id);
     return Message.ok("").data("result", result);
   }
